refactor(ui-components): moved all component type defs to correct package
