+++ src/components/ui/button.tsx
@@ -1,18 +1,8 @@
 import * as React from 'react';
 import { Slot } from '@radix-ui/react-slot';
-import { type VariantProps } from 'class-variance-authority';
 import { cn } from '@/lib/utils';
 import { buttonVariants } from '@/lib/ui/button-variants';
-
-/**
- * Button component props interface
- */
-export interface ButtonProps
-  extends React.ButtonHTMLAttributes<HTMLButtonElement>,
-    VariantProps<typeof buttonVariants> {
-  asChild?: boolean;
-  loading?: boolean;
-}
+import { type ButtonProps } from '@/types/ui/button';
 
 /**
  * Button Component
@@ -105,3 +95,4 @@ const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
 Button.displayName = 'Button';
 
 export { Button };
+export type { ButtonProps };
