+++ src/components/ui/input.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import { type VariantProps } from 'class-variance-authority';
 import { cn } from '@/lib/utils';
 import {
   inputVariants,
@@ -11,40 +10,7 @@ import {
   clearButtonVariants,
   characterCounterVariants,
 } from '@/lib/ui/input-variants';
-
-/**
- * Base Input component props interface
- */
-export interface InputProps
-  extends Omit<React.InputHTMLAttributes<HTMLInputElement>, 'size'>,
-    VariantProps<typeof inputVariants> {
-  // Input styling
-  variant?: 'default' | 'filled' | 'outlined';
-  size?: 'sm' | 'default' | 'lg';
-  state?: 'default' | 'error' | 'success' | 'warning';
-
-  // Label and text
-  label?: string;
-  helperText?: string;
-  errorText?: string;
-
-  // Icons and interactions
-  leftIcon?: React.ReactNode;
-  rightIcon?: React.ReactNode;
-  clearable?: boolean;
-  onClear?: () => void;
-
-  // Advanced features
-  loading?: boolean;
-  floatingLabel?: boolean;
-  showCharacterCount?: boolean;
-  characterLimit?: number;
-
-  // Container props
-  containerClassName?: string;
-  labelClassName?: string;
-  helperClassName?: string;
-}
+import { type InputProps } from '@/types/ui/input';
 
 /**
  * Input Component
@@ -383,3 +349,4 @@ const Input = React.forwardRef<HTMLInputElement, InputProps>(
 Input.displayName = 'Input';
 
 export { Input };
+export type { InputProps };
