+++ src/components/ui/table.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import { type VariantProps } from 'class-variance-authority';
 import { cn } from '@/lib/utils';
 import {
   tableVariants,
@@ -11,66 +10,21 @@ import {
   tableCellVariants,
   tableCaptionVariants,
 } from '@/lib/ui/table-variants';
-
-/**
- * Table component props interface
- */
-export interface TableProps
-  extends React.TableHTMLAttributes<HTMLTableElement>,
-    VariantProps<typeof tableVariants> {}
-
-/**
- * Table Header component props interface
- */
-export interface TableHeaderProps
-  extends React.HTMLAttributes<HTMLTableSectionElement>,
-    VariantProps<typeof tableHeaderVariants> {}
-
-/**
- * Table Body component props interface
- */
-export interface TableBodyProps
-  extends React.HTMLAttributes<HTMLTableSectionElement>,
-    VariantProps<typeof tableBodyVariants> {}
-
-/**
- * Table Footer component props interface
- */
-export interface TableFooterProps
-  extends React.HTMLAttributes<HTMLTableSectionElement>,
-    VariantProps<typeof tableFooterVariants> {}
-
-/**
- * Table Row component props interface
- */
-export interface TableRowProps
-  extends React.HTMLAttributes<HTMLTableRowElement>,
-    VariantProps<typeof tableRowVariants> {}
-
-/**
- * Table Head component props interface
- */
-export interface TableHeadProps
-  extends Omit<React.ThHTMLAttributes<HTMLTableCellElement>, 'align'>,
-    VariantProps<typeof tableHeadVariants> {
-  align?: 'left' | 'center' | 'right';
-}
-
-/**
- * Table Cell component props interface
- */
-export interface TableCellProps
-  extends Omit<React.TdHTMLAttributes<HTMLTableCellElement>, 'align'>,
-    VariantProps<typeof tableCellVariants> {
-  align?: 'left' | 'center' | 'right';
-}
-
-/**
- * Table Caption component props interface
- */
-export interface TableCaptionProps
-  extends React.HTMLAttributes<HTMLTableCaptionElement>,
-    VariantProps<typeof tableCaptionVariants> {}
+import {
+  type TableProps,
+  type TableHeaderProps,
+  type TableBodyProps,
+  type TableFooterProps,
+  type TableRowProps,
+  type TableHeadProps,
+  type TableCellProps,
+  type TableCaptionProps,
+  type RecipeTableProps,
+  type Ingredient,
+  type IngredientsTableProps,
+  type NutritionItem,
+  type NutritionTableProps,
+} from '@/types/ui/table';
 
 /**
  * Main table component
@@ -194,28 +148,6 @@ const TableCaption = React.forwardRef<
 TableCaption.displayName = 'TableCaption';
 
 /**
- * Recipe Table component props interface
- */
-export interface RecipeTableProps extends Omit<TableProps, 'children'> {
-  data: Array<Record<string, unknown>>;
-  columns: Array<{
-    key: string;
-    header: React.ReactNode;
-    width?: string;
-    align?: 'left' | 'center' | 'right';
-    sortable?: boolean;
-    render?: (
-      value: unknown,
-      row: Record<string, unknown>,
-      index: number
-    ) => React.ReactNode;
-  }>;
-  caption?: React.ReactNode;
-  emptyMessage?: string;
-  onRowClick?: (row: Record<string, unknown>, index: number) => void;
-}
-
-/**
  * Recipe-specific table component for displaying structured data
  */
 function RecipeTableComponent(
@@ -281,31 +213,6 @@ function RecipeTableComponent(
 const RecipeTable = React.forwardRef(RecipeTableComponent);
 
 RecipeTable.displayName = 'RecipeTable';
-
-/**
- * Ingredient interface
- */
-export interface Ingredient {
-  id?: string | number;
-  name: string;
-  amount?: number | string;
-  unit?: string;
-  notes?: string;
-  category?: string;
-  optional?: boolean;
-  [key: string]: unknown;
-}
-
-/**
- * Ingredients Table component props interface
- */
-export interface IngredientsTableProps
-  extends Omit<RecipeTableProps, 'columns' | 'data'> {
-  ingredients: Ingredient[];
-  showCategory?: boolean;
-  showNotes?: boolean;
-  onIngredientClick?: (ingredient: Ingredient, index: number) => void;
-}
 
 /**
  * Specialized table for recipe ingredients
@@ -427,28 +334,6 @@ const IngredientsTable = React.forwardRef<
 IngredientsTable.displayName = 'IngredientsTable';
 
 /**
- * Nutrition item interface
- */
-export interface NutritionItem {
-  nutrient: string;
-  amount: number | string;
-  unit: string;
-  dailyValue?: number | string;
-  category?: 'macronutrient' | 'vitamin' | 'mineral' | 'other';
-  [key: string]: unknown;
-}
-
-/**
- * Nutrition Table component props interface
- */
-export interface NutritionTableProps
-  extends Omit<RecipeTableProps, 'columns' | 'data'> {
-  nutrition: NutritionItem[];
-  showDailyValue?: boolean;
-  showCategory?: boolean;
-}
-
-/**
  * Specialized table for nutrition information
  */
 const NutritionTable = React.forwardRef<HTMLTableElement, NutritionTableProps>(
@@ -566,4 +451,20 @@ export {
   RecipeTable,
   IngredientsTable,
   NutritionTable,
+};
+
+export type {
+  TableProps,
+  TableHeaderProps,
+  TableBodyProps,
+  TableFooterProps,
+  TableRowProps,
+  TableHeadProps,
+  TableCellProps,
+  TableCaptionProps,
+  RecipeTableProps,
+  Ingredient,
+  IngredientsTableProps,
+  NutritionItem,
+  NutritionTableProps,
 };
