+++ src/types/ui/button.ts
@@ -1,0 +1,13 @@
+import * as React from 'react';
+import { type VariantProps } from 'class-variance-authority';
+import { buttonVariants } from '@/lib/ui/button-variants';
+
+/**
+ * Button component props interface
+ */
+export interface ButtonProps
+  extends React.ButtonHTMLAttributes<HTMLButtonElement>,
+    VariantProps<typeof buttonVariants> {
+  asChild?: boolean;
+  loading?: boolean;
+}
