+++ src/types/ui/card.ts
@@ -1,0 +1,50 @@
+import * as React from 'react';
+import { type VariantProps } from 'class-variance-authority';
+import { cardVariants } from '@/lib/ui/card-variants';
+
+/**
+ * Card component props interface
+ */
+export interface CardProps
+  extends React.HTMLAttributes<HTMLDivElement>,
+    VariantProps<typeof cardVariants> {
+  asChild?: boolean;
+  interactive?: boolean;
+}
+
+/**
+ * Card Header component props interface
+ */
+export interface CardHeaderProps extends React.HTMLAttributes<HTMLDivElement> {
+  asChild?: boolean;
+}
+
+/**
+ * Card Content component props interface
+ */
+export interface CardContentProps extends React.HTMLAttributes<HTMLDivElement> {
+  asChild?: boolean;
+}
+
+/**
+ * Card Footer component props interface
+ */
+export interface CardFooterProps extends React.HTMLAttributes<HTMLDivElement> {
+  asChild?: boolean;
+}
+
+/**
+ * Card Title component props interface
+ */
+export interface CardTitleProps
+  extends React.HTMLAttributes<HTMLHeadingElement> {
+  asChild?: boolean;
+}
+
+/**
+ * Card Description component props interface
+ */
+export interface CardDescriptionProps
+  extends React.HTMLAttributes<HTMLParagraphElement> {
+  asChild?: boolean;
+}
